core, all: Large buffers and FFTs work in progress
This is an overhaul of the GNURadio runtime to support
large buffers and FFT's of size greater than 2^22 bins.

The changes ripple throughout the code, and
necessarily break the API to achieve the objectives.

This is a commit to save work in progress that has
not been compile tested

Remaining TODOs to get to code that compiles:

Fix remaining code for changes in
   io_signature() item_size change from int to size_t
   history() change from int to size_t
   output_multiple() change from int to size_t
   general_/work() change to ssize_t, size_t, and gr_vector_size_t
   forecast() change to size_t and gr_vector_size_t
   noutput_items and ninput_items[] max increase from MAX_INT to MAX_SSIZE_T

As of this commit, the following parts of the tree still need to be worked:
   gr-blocks
   gr-digital
   gr-dtv
   gr-fec
   gr-qtgui
   gr-trellis
   gr-vocoder
   gr-wxgui
   gr-zeromq
   gnurado-runtime (only nco, vco, sin/cos batch calls: from int to size_t)
   volk (can only safely batch up to MAX_INT items due to ORC and asm)

Moot tasks:
FFTW limits FFTs to size 2^31-1; FFTs can be no larger.
FFTs & window functions already handle this 2^31-1 limit.
GRC looks doesn't appear to need any changes

Deferred Tasks:
Update gr_modtool (gr-utils) templates and version decision code for new API
Expanding sample delay for tags from int to size_t

+++ gnuradio-runtime/include/gnuradio/block_detail.h
@@ -74,25 +74,25 @@ namespace gr {
      * \brief Tell the scheduler \p how_many_items of input stream \p
      * which_input were consumed.
      */
-    void consume(int which_input, int how_many_items);
+    void consume(int which_input, size_t how_many_items);
 
     /*!
      * \brief Tell the scheduler \p how_many_items were consumed on
      * each input stream.
      */
-    void consume_each(int how_many_items);
+    void consume_each(size_t how_many_items);
 
     /*!
      * \brief Tell the scheduler \p how_many_items were produced on
      * output stream \p which_output.
      */
-    void produce(int which_output, int how_many_items);
+    void produce(int which_output, size_t how_many_items);
 
     /*!
      * \brief Tell the scheduler \p how_many_items were produced on
      * each output stream.
      */
-    void produce_each(int how_many_items);
+    void produce_each(size_t how_many_items);
 
     // Return the number of items read on input stream which_input
     uint64_t nitems_read(unsigned int which_input);
@@ -200,7 +200,7 @@ namespace gr {
     gr::thread::gr_thread_t thread;    // portable thread handle
 
     void start_perf_counters();
-    void stop_perf_counters(int noutput_items, int nproduced);
+    void stop_perf_counters(size_t noutput_items, size_t nproduced);
     void reset_perf_counters();
 
     // Calls to get performance counter items
@@ -232,7 +232,7 @@ namespace gr {
     float pc_work_time_total();
 
     tpb_detail d_tpb;	// used by thread-per-block scheduler
-    int d_produce_or;
+    size_t d_produce_or;
 
     // ----------------------------------------------------------------------------
