+++ gnuradio-runtime/lib/scheduler_tpb.h
@@ -42,11 +42,11 @@ namespace gr {
      * The scheduler will continue running until all blocks until they
      * report that they are done or the stop method is called.
      */
-    scheduler_tpb(flat_flowgraph_sptr ffg, int max_noutput_items);
+    scheduler_tpb(flat_flowgraph_sptr ffg, size_t max_noutput_items);
 
   public:
     static scheduler_sptr make(flat_flowgraph_sptr ffg,
-                               int max_noutput_items=100000);
+                               size_t max_noutput_items=100000);
 
     ~scheduler_tpb();
