+++ gnuradio-runtime/lib/tpb_thread_body.h
@@ -40,7 +40,7 @@ namespace gr {
     block_executor d_exec;
 
   public:
-    tpb_thread_body(block_sptr block, int max_noutput_items=100000);
+    tpb_thread_body(block_sptr block, size_t max_noutput_items=100000);
     ~tpb_thread_body();
   };
