+++ gnuradio-runtime/lib/vmcircbuf_mmap_tmpfile.h
@@ -35,7 +35,7 @@ namespace gr {
   class GR_RUNTIME_API vmcircbuf_mmap_tmpfile : public gr::vmcircbuf
   {
   public:
-    vmcircbuf_mmap_tmpfile(int size);
+    vmcircbuf_mmap_tmpfile(size_t size);
     virtual ~vmcircbuf_mmap_tmpfile();
   };
 
@@ -62,7 +62,7 @@ namespace gr {
      *
      * Call this to create a doubly mapped circular buffer.
      */
-    virtual gr::vmcircbuf *make(int size);
+    virtual gr::vmcircbuf *make(size_t size);
   };
 
 } /* namespace gr */
