+++ gr-analog/include/gnuradio/analog/agc.h
@@ -78,9 +78,9 @@ namespace gr {
 	  return output;
 	}
 
-	void scaleN(gr_complex output[], const gr_complex input[], unsigned n)
+	void scaleN(gr_complex output[], const gr_complex input[], size_t n)
 	{
-	  for(unsigned i = 0; i < n; i++) {
+	  for(size_t i = 0; i < n; i++) {
 	    output[i] = scale (input[i]);
 	  }
 	}
@@ -134,9 +134,9 @@ namespace gr {
 	  return output;
 	}
 
-	void scaleN(float output[], const float input[], unsigned n)
+	void scaleN(float output[], const float input[], size_t n)
 	{
-	  for(unsigned i = 0; i < n; i++)
+	  for(size_t i = 0; i < n; i++)
 	    output[i] = scale (input[i]);
 	}
