+++ gr-analog/lib/agc_cc_impl.cc
@@ -53,15 +53,15 @@ namespace gr {
     {
     }
 
-    int
-    agc_cc_impl::work(int noutput_items,
+    ssize_t
+    agc_cc_impl::work(size_t noutput_items,
 		      gr_vector_const_void_star &input_items,
 		      gr_vector_void_star &output_items)
     {
       const gr_complex *in = (const gr_complex*)input_items[0];
       gr_complex *out = (gr_complex*)output_items[0];
       scaleN(out, in, noutput_items);
-      return noutput_items;
+      return static_cast<ssize_t>(noutput_items);
     }
 
   } /* namespace analog */
